backend/requirements.txt
# Flask Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1

# API Clients
requests==2.31.0
google-generativeai==0.3.2

# Utilities
python-dotenv==1.0.0
pyarrow==15.0.0
